requirements.txt
python-dotenv~=1.0.0
aiogram~=3.0.0b8
loguru~=0.7.0
aiohttp~=3.8.4
SQLAlchemy~=2.0.19